% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dualGSEA}
\alias{dualGSEA}
\title{Reimplementation of dualGSEA (Bull et al., 2024) but defaults with
replaid backend. For the preranked test we still use fgsea. Should
be much faster than original using fgsea + GSVA::ssGSEA.}
\usage{
dualGSEA(
  X,
  y,
  gmt,
  matG = NULL,
  fc.test = "fgsea",
  gsea.method = "replaid.ssgsea"
)
}
\description{
Reimplementation of dualGSEA (Bull et al., 2024) but defaults with
replaid backend. For the preranked test we still use fgsea. Should
be much faster than original using fgsea + GSVA::ssGSEA.
}
