% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fc_ztest}
\alias{fc_ztest}
\title{Z-test statistical testing of differentially enrichment}
\usage{
fc_ztest(fc, G, zmat = FALSE, alpha = 0.5)
}
\arguments{
\item{fc}{Vector of logFC values}

\item{G}{Sparse matrix of gene sets. Non-zero entry indicates
gene/feature is part of gene sets. Features on rows, gene sets on
columns.}

\item{zmat}{Logical indicating to return z-matrix}

\item{alpha}{Scalar weight for SD estimation. Default 0.5.}
}
\value{
List with element: z_statistic (z-statistic from one-sample z-test),
p_value (p-value from one-sample z-test), and zmat (z-matrix).
}
\description{
This function performs statistical testing for differential
enrichment using plaid
}
