% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gset_ttest}
\alias{gset_ttest}
\title{Perform t-test on gene set scores}
\usage{
gset_ttest(gsetX, y)
}
\arguments{
\item{gsetX}{Matrix of gene set scores with gene sets on rows and samples on columns}

\item{y}{Binary vector (0/1) indicating group membership}
}
\value{
Data frame with columns: diff (difference in means), statistic (t-statistic),
pvalue (p-value), and other t-test results.
}
\description{
Perform t-test on gene set scores
}
