% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{mat.rowsds}
\alias{mat.rowsds}
\title{Calculate row standard deviations for matrix}
\usage{
mat.rowsds(X)
}
\arguments{
\item{X}{Input matrix (can be sparse or dense)}
}
\value{
Vector of row standard deviations.
}
\description{
Calculate row standard deviations for matrix
}
