% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{matrix_metap}
\alias{matrix_metap}
\title{Matrix version for combining p-values using fisher or stouffer
method. Much faster than doing metap::sumlog() and metap::sumz()}
\usage{
matrix_metap(plist, method = "stouffer")
}
\description{
Matrix version for combining p-values using fisher or stouffer
method. Much faster than doing metap::sumlog() and metap::sumz()
}
