% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{matrix_onesample_ttest}
\alias{matrix_onesample_ttest}
\title{Perform one-sample t-test on matrix with gene sets}
\usage{
matrix_onesample_ttest(Fm, G)
}
\arguments{
\item{Fm}{Vector of feature values (e.g., fold changes)}

\item{G}{Sparse matrix of gene sets with genes on rows and gene sets on columns}
}
\value{
List containing mean, t-statistic, and p-value matrices.
}
\description{
Perform one-sample t-test on matrix with gene sets
}
