% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{plaid.test}
\alias{plaid.test}
\title{Statistical testing of differentially enrichment}
\usage{
plaid.test(X, y, G, gsetX, tests = c("one", "lm"), sort.by = "p.meta")
}
\arguments{
\item{X}{Matrix of log expression value}

\item{y}{Vector of 0s and 1s indicating group}

\item{G}{Sparse matrix of gene sets. Non-zero entry indicates
gene/feature is part of gene sets. Features on rows, gene sets on
columns.}

\item{gsetX}{Gene set score matrix which is output of
\code{plaid()}. Can be NULL in that case it will be recomputed from X
and G (default required).}

\item{tests}{Character array indicating which tests to perform.}
}
\description{
This function performs statistical testing for differential
enrichment using plaid
}
