% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{replaid.aucell}
\alias{replaid.aucell}
\title{Fast calculation of AUCell}
\usage{
replaid.aucell(X, matG, aucMaxRank = ceiling(0.05 * nrow(X)))
}
\arguments{
\item{X}{Gene or protein expression matrix. Generally log
transformed. See details. Genes on rows, samples on columns.}

\item{matG}{Gene sets sparse matrix. Genes on rows, gene sets on columns.}

\item{aucMaxRank}{Rank threshold (see AUCell paper). Default aucMaxRank = 0.05*nrow(X).}
}
\description{
Calculates single-sample enrichment AUCell (Aibar
et al., 2017) using plaid back-end. The computation is
10-100x faster than the original code.
}
\details{
Computing the AUCell score requires to compute the ranks
of the expression matrix and approximating the AUC of a gene
set. We have wrapped this in a single convenience function.

We have extensively compared the results of \code{replaid.aucell} and
from the original \code{UCell} R package and we showed good concordance
of results in the score, logFC and p-values.
}
