% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{sparse_colranks}
\alias{sparse_colranks}
\title{Compute columm ranks for sparse matrix. Internally used by colranks()}
\usage{
sparse_colranks(X, signed = FALSE, ties.method = "average")
}
\arguments{
\item{X}{Input matrix}

\item{signed}{Logical: use or not signed ranks}

\item{ties.method}{Character Choice of ties.method}
}
\value{
Sparse matrix of columnwise ranks with same dimensions as input.
}
\description{
Compute columm ranks for sparse matrix. Internally used by colranks()
}
