% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{chunked_crossprod}
\alias{chunked_crossprod}
\title{Compute crossprod (t(x) \%*\% y) for very large y by computing in
chunks.}
\usage{
chunked_crossprod(x, y, chunk = NULL)
}
\description{
Compute crossprod (t(x) \%*\% y) for very large y by computing in
chunks.
}
