% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{colranks}
\alias{colranks}
\title{Compute columnwise ranks of matrix}
\usage{
colranks(
  X,
  sparse = NULL,
  signed = FALSE,
  keep.zero = FALSE,
  ties.method = "average"
)
}
\arguments{
\item{X}{Input matrix}

\item{sparse}{Logical indicating to use sparse methods}

\item{signed}{Logical indicating using signed ranks}

\item{keep.zero}{Logical indicating whether to keep zero as ranked zero}

\item{ties.method}{Character Choice of ties.method}
}
\description{
Computes columnwise rank of matrix. Can be sparse. Tries to call
optimized functions from Rfast or matrixStats.
}
