% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dual_test}
\alias{dual_test}
\title{Statistical testing of differentially enrichment}
\usage{
dual_test(
  X,
  y,
  G,
  gsetX = NULL,
  fc.test = "cor",
  pv1 = NULL,
  pv2 = NULL,
  metap.method = "stouffer",
  sort.by = "p.dual"
)
}
\arguments{
\item{X}{Matrix of log expression value}

\item{y}{Vector of 0s and 1s indicating group}

\item{G}{Sparse matrix of gene sets. Non-zero entry indicates
gene/feature is part of gene sets. Features on rows, gene sets on
columns.}

\item{gsetX}{Gene set score matrix which is output of
\code{plaid()}. Can be NULL in that case it will be recomputed from X
and G (default required).}

\item{tests}{Character array indicating which tests to perform.}
}
\description{
This function performs statistical testing for differential
enrichment using plaid
}
