% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fc_ttest}
\alias{fc_ttest}
\title{Statistical testing of differentially enrichment}
\usage{
fc_ttest(fc, G, sort.by = "pvalue")
}
\arguments{
\item{fc}{Vector of logFC values}

\item{G}{Sparse matrix of gene sets. Non-zero entry indicates
gene/feature is part of gene sets. Features on rows, gene sets on
columns.}
}
\description{
This function performs statistical testing for differential
enrichment using plaid
}
