% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gset_averageCLR}
\alias{gset_averageCLR}
\title{Compute geneset expression as the average log-ration of genes in
the geneset. Requires log-expression matrix X and (sparse) geneset
matrix matG.}
\usage{
gset_averageCLR(X, matG, center = TRUE)
}
\arguments{
\item{X}{Log-expression matrix with genes on rows and samples on columns}

\item{matG}{Sparse gene set matrix with genes on rows and gene sets on columns}

\item{center}{Logical indicating whether to center the results}
}
\value{
Matrix of gene set expression scores with gene sets on rows and samples on columns.
}
\description{
Compute geneset expression as the average log-ration of genes in
the geneset. Requires log-expression matrix X and (sparse) geneset
matrix matG.
}
