% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gset_averageCLR}
\alias{gset_averageCLR}
\title{Compute geneset expression as the average log-ration of genes in
the geneset. Requires log-expression matrix X and (sparse) geneset
matrix matG.}
\usage{
gset_averageCLR(X, matG, center = TRUE, use.rank = FALSE)
}
\description{
Compute geneset expression as the average log-ration of genes in
the geneset. Requires log-expression matrix X and (sparse) geneset
matrix matG.
}
