% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{normalize_medians}
\alias{normalize_medians}
\title{Normalize column medians of matrix}
\usage{
normalize_medians(x, ignore.zero = NULL)
}
\arguments{
\item{x}{Input matrix}

\item{ignore.zero}{Logical indicating whether to ignore zeros to
exclude for median calculation}
}
\description{
This function normalizes the column medians of matrix x. It calls
optimized functions from the matrixStats package.
}
