% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{plaid}
\alias{plaid}
\title{Compute geneset expression as the average log-ration of genes in
the geneset. Requires log-expression matrix X and (sparse) geneset
matrix matG.}
\usage{
plaid(X, matG, stats = "mean", chunk = NULL, normalize = TRUE)
}
\arguments{
\item{use.cov}{Logical for whether using a scaled model matrix of the full pairs}
}
\description{
Compute geneset expression as the average log-ration of genes in
the geneset. Requires log-expression matrix X and (sparse) geneset
matrix matG.
}
\examples{
# Load the example dataset provided and run:
#gmt <- read.gmt("hallmark.gmt")
#matG <- gmt2mat(gmt)
#gsetX <- plaid(X, matG)

}
