% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{replaid.gsva}
\alias{replaid.gsva}
\title{Fast approximation of GSVA}
\usage{
replaid.gsva(X, matG, tau = 0)
}
\arguments{
\item{X}{Gene or protein expression matrix. Generally log
transformed. See details. Genes on rows, samples on columns.}

\item{matG}{Gene sets sparse matrix. Genes on rows, gene sets on
columns.}

\item{tau}{Rank weight parameter (see GSVA publication). Default
tau=0.}
}
\description{
Calculates single-sample enrichment GSVA (Hänzelmann
et al., 2013) using plaid back-end. The computation is
10-100x faster than the original code.
}
\details{
Computing the GSVA score requires to compute the CDF of
the expression matrix, ranking and scoring the genesets. We have
wrapped this in a single convenience function.

We have extensively compared the results of \code{replaid.gsva} and
from the original \code{GSVA} R package and we showed good concordance
of results in the score, logFC and p-values.

In the original formulation, GSVA uses an emperical CDF to
transform expression of each feature to a \link{0;1} relative expression
value. For efficiency reasons, this is here approximated by a
z-transform (center+scale) of each row.
}
