% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{replaid.scse}
\alias{replaid.scse}
\title{Fast calculation of scSE score}
\usage{
replaid.scse(X, matG, removeLog2 = NULL, scoreMean = FALSE)
}
\arguments{
\item{X}{Gene or protein expression matrix. Generally log
transformed. See details. Genes on rows, samples on columns.}

\item{matG}{Gene sets sparse matrix. Genes on rows, gene sets on
columns.}

\item{removeLog2}{Logical for whether to remove the Log2, i.e. will
apply power transform (base2) on input (default TRUE).}

\item{scoreMean}{Logical for whether computing sum or mean as score
(default FALSE).}
}
\description{
Calculates Single-Cell Signature Explorer (Pont et
al., 2019) scores using plaid back-end. The computation is
10-100x faster than the original code.
}
\details{
Computing the scSE requires running plaid on the linear
(not logarithmic) score and perform additional normalization by
the total UMI per sample. We have wrapped this in a single
convenience function:

To replicate the original "sum-of-UMI" scSE score, set \code{removeLog2=TRUE}
and \code{scoreMean=FALSE}. scSE and plaid scores become more similar for
\code{removeLog2=FALSE} and \code{scoreMean=TRUE}.

We have extensively compared the results from \code{replaid.scse} and
from the original scSE (implemented in GO lang) and we showed
almost identical results in the score, logFC and p-values.
}
