% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{replaid.sing}
\alias{replaid.sing}
\title{Fast calculation of singscore}
\usage{
replaid.sing(X, matG)
}
\arguments{
\item{X}{Gene or protein expression matrix. Generally log
transformed. See details. Genes on rows, samples on columns.}

\item{matG}{Gene sets sparse matrix. Genes on rows, gene sets on
columns.}

\item{removeLog2}{Logical for whether to remove the Log2, i.e. will
apply power transform (base2) on input (default TRUE).}

\item{scoreMean}{Logical for whether computing sum or mean as score
(default FALSE).}
}
\description{
Calculates single-sample enrichment singscore
(Fouratan et al., 2018) using plaid back-end. The computation is
10-100x faster than the original code.
}
\details{
Computing the singscore requires to compute the ranks of
the expression matrix. We have wrapped this in a single
convenience function.

We have extensively compared the results of \code{replaid.sing} and from
the original \code{singscore} R package and we showed identical result
in the score, logFC and p-values.
}
