% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{replaid.ssgsea}
\alias{replaid.ssgsea}
\title{Fast calculation of ssGSEA}
\usage{
replaid.ssgsea(X, matG, alpha = 0)
}
\arguments{
\item{X}{Gene or protein expression matrix. Generally log
transformed. See details. Genes on rows, samples on columns.}

\item{matG}{Gene sets sparse matrix. Genes on rows, gene sets on
columns.}

\item{alpha}{Weighting factor for exponential weighting of ranks}
}
\description{
Calculates single-sample enrichment singscore (Barbie
et al., 2009; Hänzelmann et al., 2013) using plaid back-end. The
computation is 10-100x faster than the original code.
}
\details{
Computing ssGSEA score requires to compute the ranks of
the expression matrix and weighting of the ranks. We have wrapped
this in a single convenience function.

We have extensively compared the results of \code{replaid.ssgsea} and
from the original \code{GSVA} R package and we showed highly similar
results in the score, logFC and p-values. For alpha=0 we obtain
exact results, for alpha>0 the results are highly similar but not
exactly the same.
}
