% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{replaid.ucell}
\alias{replaid.ucell}
\title{Fast calculation of UCell}
\usage{
replaid.ucell(X, matG, rmax = 1500)
}
\arguments{
\item{X}{Gene or protein expression matrix. Generally log
transformed. See details. Genes on rows, samples on columns.}

\item{matG}{Gene sets sparse matrix. Genes on rows, gene sets on
columns.}

\item{rmax}{Rank truncation threshold (see original
publication). Default rmax=1500.}
}
\description{
Calculates single-sample enrichment UCell (Andreatta
et al., 2021) using plaid back-end. The computation is
10-100x faster than the original code.
}
\details{
Computing ssGSEA score requires to compute the ranks of
the expression matrix and truncation of the ranks. We have wrapped
this in a single convenience function.

We have extensively compared the results of \code{replaid.ucell} and
from the original \code{UCell} R package and we showed near exacct
results in the score, logFC and p-values.
}
