% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaid.R
\name{sparse_colranks}
\alias{sparse_colranks}
\title{Test if rows or columns of sparse matrix are duplicated}
\usage{
sparse_colranks(X, signed = FALSE, ties.method = "average")
}
\description{
Test if rows or columns of sparse matrix are duplicated
}
